Move md5 to stdlib nifs from emscripten platform nifs
